- Added the OASIS logo into a new solution folder Images/Logos so all projects can then refrence this for their NuGet packages and so will never have issue finding the correct logo/image again.
- Fixed the NextGenSoftware.OASIS.API.ONODE.Core.TestHarness project.

- Fixed a bug in the get-all-avatar-names-grouped-by-name avatar api function.

- Changed return type from OASISResult<int> to OASISResult<long> for GetKarmaForAvatar function in KarmaController in OASIS.API.ONODE.WebApi.

- NextGenSoftware.OASIS.OASISBootLoader now references new Images/Logo/OASISLogo128.jpg.

+++ NextGenSoftware.OASIS.API.ONODE.WebAPI/Controllers/KarmaController.cs
@@ -75,14 +75,14 @@ namespace NextGenSoftware.OASIS.API.ONode.WebAPI.Controllers
         /// <param name="setGlobally"> Set this to false for this provider to be used only for this request or true for it to be used for all future requests too.</param>
         /// <returns></returns>
         [HttpGet("get-karma-for-avatar/{avatarId}")]
-        public OASISResult<int> GetKarmaForAvatar(Guid avatarId)
+        public OASISResult<long> GetKarmaForAvatar(Guid avatarId)
         {
             OASISResult<IAvatarDetail> avatarDetailResult = Program.AvatarManager.LoadAvatarDetail(avatarId);
 
             if (!avatarDetailResult.IsError && avatarDetailResult.Result != null)
-                return new OASISResult<int>(avatarDetailResult.Result.Karma);
+                return new OASISResult<long>(avatarDetailResult.Result.Karma);
             else
-                return new OASISResult<int>() { IsError = true, Message = $"Error loading avatar detail. Reason:{avatarDetailResult.Message}" };
+                return new OASISResult<long>() { IsError = true, Message = $"Error loading avatar detail. Reason:{avatarDetailResult.Message}" };
         }
 
         /// <summary>
@@ -93,7 +93,7 @@ namespace NextGenSoftware.OASIS.API.ONode.WebAPI.Controllers
         /// <param name="setGlobally"> Set this to false for this provider to be used only for this request or true for it to be used for all future requests too.</param>
         /// <returns></returns>
         [HttpGet("get-karma-for-avatar/{avatarId}/{providerType}/{setGlobally}")]
-        public OASISResult<int> GetKarmaForAvatar(Guid avatarId, ProviderType providerType, bool setGlobally = false)
+        public OASISResult<long> GetKarmaForAvatar(Guid avatarId, ProviderType providerType, bool setGlobally = false)
         {
             GetAndActivateProvider(providerType, setGlobally);
             return GetKarmaForAvatar(avatarId);
